common fallback test
