+++ stats/stats/src/charts/counters/total_blocks.rs
@@ -103,14 +103,15 @@ mod tests {
     use crate::{
         data_source::{types::BlockscoutMigrations, DataSource, UpdateContext, UpdateParameters},
         tests::{
-            init_db::init_marked_db_all, mock_blockscout::fill_mock_blockscout_data,
-            simple_test::get_counter,
+            init_db::init_marked_db_all,
+            mock_blockscout::fill_mock_blockscout_data,
+            simple_test::{get_counter, test_counter_fallback},
         },
     };
     use chrono::NaiveDate;
     use entity::chart_data;
-    use pretty_assertions::{assert_eq, assert_ne};
-    use sea_orm::{DatabaseConnection, DbBackend, EntityTrait, Set, Statement};
+    use pretty_assertions::assert_eq;
+    use sea_orm::{DatabaseConnection, EntityTrait, Set};
     use std::str::FromStr;
 
     #[tokio::test]
@@ -217,34 +218,6 @@ mod tests {
     #[tokio::test]
     #[ignore = "needs database to run"]
     async fn total_blocks_fallback() {
-        let _ = tracing_subscriber::fmt::try_init();
-        let (db, blockscout) = init_marked_db_all("total_blocks_fallback").await;
-        let current_time = chrono::DateTime::from_str("2023-03-01T12:00:00Z").unwrap();
-        let current_date = current_time.date_naive();
-
-        TotalBlocks::init_recursively(&db.connection, &current_time)
-            .await
-            .unwrap();
-
-        fill_mock_blockscout_data(&blockscout.connection, current_date).await;
-
-        // need to analyze or vacuum for `reltuples` to be updated.
-        // source: https://www.postgresql.org/docs/9.3/planner-stats.html
-        let _ = blockscout
-            .connection
-            .execute(Statement::from_string(DbBackend::Postgres, "ANALYZE;"))
-            .await
-            .unwrap();
-
-        let parameters = UpdateParameters {
-            db: &db,
-            blockscout: &blockscout,
-            blockscout_applied_migrations: BlockscoutMigrations::latest(),
-            update_time_override: Some(current_time),
-            force_full: false,
-        };
-        let cx: UpdateContext<'_> = UpdateContext::from_params_now_or_override(parameters.clone());
-        let data = get_counter::<TotalBlocks>(&cx).await;
-        assert_ne!("0", data.value);
+        test_counter_fallback::<TotalBlocks>("total_blocks_fallback").await;
     }
 }
